Lesson 7

+++ routes/web.php
@@ -24,34 +24,39 @@ use App\Models\Listing;
 Route::get('/', [ListingController::class, 'index']);
 
 // Show edit form
-Route::get('/listings/edit/{listing}', [ListingController::class, 'edit']);
+Route::get('/listings/edit/{listing}', [ListingController::class, 'edit'])->middleware('auth');
 
 // Show create form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])
+    ->middleware('auth');
 
 // Store listing data
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 // Edit submit to update
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // Edit submit to update
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 // Single listing
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 // Show Register/Create form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // Create new user
 Route::post('/users', [UserController::class, 'store']);
 
 // Create new user
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 // Show Register/Create form
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')
+    ->middleware('guest');
+
+// Login user
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
 
 // Route::get('/component', function () {
 //     return view('testcomponent');
